requirements.txt
beautifulsoup4==4.6.3
boto3==1.9.29
celery==4.2.1
Django==2.1.3
django-allauth==0.38.0
django-autoslug==1.9.3
django-bootstrap3==11.0.0
django-debug-toolbar==1.10.1
django-filter==1.0.4
django-flat-theme==1.1.3
django-jfu==2.0.9
django-nose==1.4.6
django-storages==1.7.1
django-summernote==0.8.10
django-test-without-migrations==0.6
djangorestframework==3.9.0
dj-inmemorystorage==1.4.1
psycopg2==2.7.5
nameko==2.11.0
Pillow==5.3.0
pymarc==3.1.10
python-docx==0.8.7
python-memcached==1.59
raven==6.9.0
requests==2.20.0
six==1.11.0
kombu==4.2.0